void(): adsr improved, bugfixes, linked volume slider & audio button with backend, visualisation for piano, ...

+++ dsp/dsp_utils.c
@@ -3,7 +3,7 @@
 
 // midi to freq
 float midi_to_freq(int midiNote) {
-  return 440.0f * powf(2.0f, (midiNote - 69) / 12.0f);
+  return 440.0f * powf(2.0f, ((float)midiNote - 69.0f) / 12.0f);
 }
 
 // prevent clipping
